Create longestcommonsubsequence.cpp
Using Bottom-Up DP (Tabulation) - O(m * n) Time and O(m * n) Space
There are two parameters that change in the recursive solution and these parameters go from 0 to m and 0 to n. So we create a 2D dp array of size (m+1) x (n+1).  

We first fill the known entries when m is 0 or n is 0.
Then we fill the remaining entries using the recursive formula.